[UI] Phase 4: move environment UI to ui/components/environments (#18750)
Relocates the environment surfaces from the legacy `Lifecycle/Environments/`
umbrella to a direct top-level home at `ui/components/environments/`,
matching the target Phase 4 layout in `ui/restructure-plan.md` (move
order item 7) and aligning with the workspaces/icons consolidations.

- `git mv ui/components/Lifecycle/Environments` -> `ui/components/environments`
- Rewrite relative imports inside `environment-card.tsx` and `index.tsx`
  to reflect the shallower path (one less `../`), pointing
  `Lifecycle/General` exports through `../Lifecycle/General` and
  shared `components/*` paths through `../`.
- Update the `Lifecycle/` barrel re-export to import
  `EnvironmentComponent` from `../environments` so the existing
  `pages/management/environments.tsx` and `General/Modals/EnvironmentModal.tsx`
  call sites keep working without changes.
- Refresh the eslint legacy allowlists (restricted imports, literal color,
  soft max-line) so the offender paths track the new location.

Pure relocation; no behavior or styling changes. `npm run lint`,
`npm run audit:mui`, and `npx tsc --noEmit` are clean.

Fixes #18750
Refs #18659, #18654
