+++ ui/components/Lifecycle/index.tsx
@@ -1,4 +1,4 @@
 import WorkspacesComponent from '../workspaces';
-import EnvironmentComponent from './Environments';
+import EnvironmentComponent from '../environments';
 
 export { WorkspacesComponent, EnvironmentComponent };
